SQUASH: staging: vc04_services: Fix warnings
create_pagelist uses a number of WARNs to sanity-check check the input
user pages, verifying that only the first and last entries in the
array are not integral pages. These WARNs use the output array index
when it should be input array index. Although the difference is only
significant in one instance, change all the indices for consistency.

See: https://github.com/raspberrypi/linux/pull/1987

Signed-off-by: Phil Elwell <phil@raspberrypi.org>